+++ internal/migrations/00004_create_donations_table.sql
@@ -1,9 +1,9 @@
 -- +goose Up
 CREATE TABLE donations (
-    id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE SET NULL, -- если юзер удалится, инфа о деньгах останется
-    campaign_id INTEGER NOT NULL REFERENCES campaigns(id) ON DELETE CASCADE,
-    amount DECIMAL(15, 2) NOT NULL CHECK (amount > 0),
+    id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE SET NULL, -- если юзер удалится, инфа о деньгах останется
+    campaign_id BIGINT NOT NULL REFERENCES campaigns(id) ON DELETE CASCADE,
+    amount DECIMAL(18,2) NOT NULL CHECK (amount > 0),
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
